Rename `domain` module to `domains` for consistency
