+++ cli/src/api/domains.rs
