+++ cli/src/api/domains/create.rs
